+++ src/kodo_fulcrum/fulcrum_encoder.hpp
@@ -5,27 +5,25 @@
 
 #pragma once
 
-#include <kodo/common_encoder_layers.hpp>
-#include <kodo/nested_payload_size.hpp>
-#include <kodo/nested_systematic.hpp>
-#include <kodo/nested_write_payload.hpp>
-#include <kodo/payload_precoder.hpp>
-#include <kodo/select_storage_type.hpp>
-#include <kodo/systematic_coefficient_mapper.hpp>
-#include <kodo/systematic_precoder.hpp>
-#include <kodo/trace_layer.hpp>
-#include <kodo/trace_nested_stack.hpp>
-#include <kodo/trace_systematic_coefficient_mapper.hpp>
-#include <kodo/uniform_generator_layers.hpp>
+#include <kodo_core/common_encoder_layers.hpp>
+#include <kodo_core/nested_payload_size.hpp>
+#include <kodo_core/nested_systematic.hpp>
+#include <kodo_core/nested_write_payload.hpp>
+#include <kodo_core/payload_precoder.hpp>
+#include <kodo_core/select_storage_type.hpp>
+#include <kodo_core/systematic_coefficient_mapper.hpp>
+#include <kodo_core/systematic_precoder.hpp>
+#include <kodo_core/trace_layer.hpp>
+#include <kodo_core/trace_nested_stack.hpp>
+#include <kodo_core/trace_systematic_coefficient_mapper.hpp>
+#include <kodo_core/uniform_generator_layers.hpp>
 
 #include <kodo_rlnc/shallow_full_vector_encoder.hpp>
 
-#include "fulcrum_nested_stack.hpp"
 #include "fulcrum_info.hpp"
+#include "fulcrum_nested_stack.hpp"
 
-namespace kodo
-{
-namespace fulcrum
+namespace kodo_fulcrum
 {
     /// @ingroup fec_stacks
     ///
@@ -43,39 +41,39 @@ namespace fulcrum
     >
     class fulcrum_encoder : public
         // Payload Codec API
-        nested_systematic<
-        payload_precoder<
-        systematic_precoder<
-        trace_systematic_coefficient_mapper<find_enable_trace<Features>,
-        systematic_coefficient_mapper<
-        nested_write_payload<
-        nested_payload_size<
-        trace_nested_stack<find_enable_trace<Features>,
+        kodo_core::nested_systematic<
+        kodo_core::payload_precoder<
+        kodo_core::systematic_precoder<
+        kodo_core::trace_systematic_coefficient_mapper<
+            kodo_core::find_enable_trace<Features>,
+        kodo_core::systematic_coefficient_mapper<
+        kodo_core::nested_write_payload<
+        kodo_core::nested_payload_size<
+        kodo_core::trace_nested_stack<kodo_core::find_enable_trace<Features>,
         fulcrum_nested_stack<
-            rlnc::shallow_full_vector_encoder<fifi::binary, Features>,
+            kodo_rlnc::shallow_full_vector_encoder<fifi::binary, Features>,
         // Coefficient Generator API
-        uniform_generator_layers<
+        kodo_core::uniform_generator_layers<
         // Codec API
-        common_encoder_layers<Features,
+        kodo_core::common_encoder_layers<Features,
         // Coefficient Storage API
-        coefficient_value_access<
-        coefficient_info<
+        kodo_core::coefficient_value_access<
+        kodo_core::coefficient_info<
         // Symbol Storage API
-        select_storage_type<Features,
+        kodo_core::select_storage_type<Features,
         // Finite Field API
-        finite_field_layers<Field,
+        kodo_core::finite_field_layers<Field,
         // Fulcrum API
         fulcrum_info<
             std::integral_constant<uint32_t,10>, // MaxExpansion
             std::integral_constant<uint32_t,4>,  // DefaultExpansion
         // Trace Layer
-        trace_layer<find_enable_trace<Features>,
+        kodo_core::trace_layer<kodo_core::find_enable_trace<Features>,
         // Final Layer
-        final_layer
+        kodo_core::final_layer
         >>>>>>>>>>>>>>>>>
     {
     public:
-        using factory = pool_factory<fulcrum_encoder>;
+        using factory = kodo_core::pool_factory<fulcrum_encoder>;
     };
-}
 }