+++ test/src/api/test_fulcrum_binding.cpp
@@ -23,7 +23,7 @@ namespace
         stub::call<dummy_stack*()> stack;
     };
 
-    using stack = kodo::fulcrum::api::fulcrum_binding<dummy>;
+    using stack = kodo_fulcrum::api::fulcrum_binding<dummy>;
 }
 
 TEST(api_test_fulcrum_binding, expansion)