+++ test/src/test_fulcrum_nested_inner_decoder.cpp
@@ -9,40 +9,40 @@
 
 #include <stub/call.hpp>
 
-namespace kodo
+namespace kodo_fulcrum
 {
-// Put dummy layers and tests classes in an anonymous namespace
-// to avoid violations of ODF (one-definition-rule) in other
-// translation units
-namespace
-{
-    struct nested_stack
+    // Put dummy layers and tests classes in an anonymous namespace
+    // to avoid violations of ODF (one-definition-rule) in other
+    // translation units
+    namespace
     {
-        stub::call<bool()> is_complete;
-        stub::call<uint32_t()> symbols_uncoded;
-        stub::call<bool(uint32_t)> is_symbol_uncoded;
-        stub::call<uint32_t()> rank;
-    };
+        struct nested_stack
+        {
+            stub::call<bool()> is_complete;
+            stub::call<uint32_t()> symbols_uncoded;
+            stub::call<bool(uint32_t)> is_symbol_uncoded;
+            stub::call<uint32_t()> rank;
+        };
 
-    struct dummy_layer
-    {
-        stub::call<nested_stack*()> nested;
-        stub::call<uint32_t()> symbols;
-        stub::call<uint32_t()> expansion;
+        struct dummy_layer
+        {
+            stub::call<nested_stack*()> nested;
+            stub::call<uint32_t()> symbols;
+            stub::call<uint32_t()> expansion;
 
-        nested_stack m_nested_stack;
-    };
+            nested_stack m_nested_stack;
+        };
 
-    class dummy_stack
-        : public fulcrum::fulcrum_nested_inner_decoder<dummy_layer>
-    { };
-}
+        class dummy_stack :
+            public fulcrum_nested_inner_decoder<dummy_layer>
+        { };
+    }
 }
 
 // Test that if the nested is complete we are also complete
 TEST(test_fulcrum_nested_inner_decoder, nested_is_complete)
 {
-    kodo::dummy_stack stack;
+    kodo_fulcrum::dummy_stack stack;
     stack.nested.set_return(&stack.m_nested_stack);
 
     stack.m_nested_stack.is_complete.set_return(true);
@@ -53,7 +53,7 @@ TEST(test_fulcrum_nested_inner_decoder, nested_is_complete)
 // outer code symbols
 TEST(test_fulcrum_nested_inner_decoder, uncoded_less_than_symbols)
 {
-    kodo::dummy_stack stack;
+    kodo_fulcrum::dummy_stack stack;
     stack.nested.set_return(&stack.m_nested_stack);
 
     stack.m_nested_stack.is_complete.set_return(false);
@@ -66,7 +66,7 @@ TEST(test_fulcrum_nested_inner_decoder, uncoded_less_than_symbols)
 // User part of the data is decoded
 TEST(test_fulcrum_nested_inner_decoder, user_decoded)
 {
-    kodo::dummy_stack stack;
+    kodo_fulcrum::dummy_stack stack;
     stack.nested.set_return(&stack.m_nested_stack);
 
     stack.m_nested_stack.is_complete.set_return(false);
@@ -80,7 +80,7 @@ TEST(test_fulcrum_nested_inner_decoder, user_decoded)
 // User part of the data is not decoded
 TEST(test_fulcrum_nested_inner_decoder, user_not_decoded)
 {
-    kodo::dummy_stack stack;
+    kodo_fulcrum::dummy_stack stack;
     stack.nested.set_return(&stack.m_nested_stack);
 
     stack.m_nested_stack.is_complete.set_return(false);
@@ -96,7 +96,7 @@ TEST(test_fulcrum_nested_inner_decoder, user_not_decoded)
 // difference
 TEST(test_fulcrum_rank_inner_decoder, rank_zero)
 {
-    kodo::dummy_stack stack;
+    kodo_fulcrum::dummy_stack stack;
     stack.nested.set_return(&stack.m_nested_stack);
     stack.expansion.set_return(4U);
     stack.m_nested_stack.rank.set_return({3U, 4U, 5U, 6U});