+++ test/src/test_fulcrum_nested_stack.cpp
@@ -3,11 +3,12 @@
 // See accompanying file LICENSE.rst or
 // http://www.steinwurf.com/licensing
 
-#include <gtest/gtest.h>
-
 #include <kodo_fulcrum/fulcrum_nested_stack.hpp>
 
-namespace kodo
+#include <gtest/gtest.h>
+
+
+namespace kodo_fulcrum
 {
     // Put dummy layers and tests classes in an anonymous namespace
     // to avoid violations of ODF (one-definition-rule) in other
@@ -127,7 +128,7 @@ namespace kodo
 
         template<uint32_t MaxExpansion>
         class dummy_stack : public
-            fulcrum::fulcrum_nested_stack<
+            fulcrum_nested_stack<
             nested_dummy_stack, dummy_layer<MaxExpansion> >
         { };
     }
@@ -139,7 +140,7 @@ TEST(test_fulcrum_nested_stack, api)
     uint32_t max_symbols = 10;
     uint32_t max_symbol_size = 10;
 
-    using test_stack = kodo::dummy_stack<max_expansion>;
+    using test_stack = kodo_fulcrum::dummy_stack<max_expansion>;
 
     test_stack::config factory(max_symbols, max_symbol_size);