+++ test/src/test_fulcrum_payload_decoder.cpp
@@ -9,7 +9,7 @@
 #include <gtest/gtest.h>
 #include <stub/call.hpp>
 
-namespace kodo
+namespace kodo_fulcrum
 {
     // Put dummy layers and tests classes in an anonymous namespace
     // to avoid violations of ODF (one-definition-rule) in other
@@ -40,14 +40,14 @@ namespace kodo
 /// dependency.
 TEST(test_fulcrum_payload_decoder, api)
 {
-    using encoder_stack = kodo::fulcrum::fulcrum_encoder<fifi::binary8>;
+    using encoder_stack = kodo_fulcrum::fulcrum_encoder<fifi::binary8>;
 
-    using payload_stack = kodo::fulcrum::fulcrum_payload_decoder<
-        kodo::dummy_main_stack>;
+    using payload_stack = kodo_fulcrum::fulcrum_payload_decoder<
+        kodo_fulcrum::dummy_main_stack>;
 
     // The main stack where the call to read_symbol(..) will go
-    kodo::dummy_main_stack::config main_factory;
-    kodo::dummy_main_stack main_stack;
+    kodo_fulcrum::dummy_main_stack::config main_factory;
+    kodo_fulcrum::dummy_main_stack main_stack;
 
     uint32_t expansion = 2U;
     uint32_t symbols = 16;