+++ NEWS.rst
@@ -6,7 +6,8 @@ detailed list of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Upgrade to kodo-rlnc 2
+* Major: Renamed the ``kodo::fulcrum`` namespace to ``kodo_fulcrum``.
 
 1.0.0
 -----